fix LOGBACK-1390

+++ logback-core/src/main/java/ch/qos/logback/core/rolling/helper/RollingCalendar.java
@@ -187,7 +187,6 @@ public class RollingCalendar extends GregorianCalendar {
         long diff = endFloored - startFloored;
         
         switch (periodicityType) {
-
         case TOP_OF_MILLISECOND:
             return diff;
         case TOP_OF_SECOND:
@@ -195,8 +194,7 @@ public class RollingCalendar extends GregorianCalendar {
         case TOP_OF_MINUTE:
             return diff / MILLIS_IN_ONE_MINUTE;
         case TOP_OF_HOUR:
-
-            return (int) diff / MILLIS_IN_ONE_HOUR;
+            return diff / MILLIS_IN_ONE_HOUR;
         case TOP_OF_DAY:
             return diff / MILLIS_IN_ONE_DAY;
         case TOP_OF_WEEK:
